Move settings into separate module
