+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "crates-lsp"
-version = "0.0.3"
+version = "0.0.4"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
