+++ src/main.rs
@@ -1,18 +1,16 @@
-use std::sync::Arc;
-
 use crates::api::CrateApi;
 use crates::cache::CrateCache;
 use crates::sparse::CrateIndex;
 use crates::CrateLookup;
 use parse::{DependencyVersion, ManifestTracker};
-use serde::Deserialize;
-use tokio::sync::RwLock;
+use settings::Settings;
 use tower_lsp::jsonrpc::Result;
 use tower_lsp::lsp_types::*;
 use tower_lsp::{Client, LanguageServer, LspService, Server};
 
 mod crates;
 mod parse;
+mod settings;
 
 #[derive(Debug, Clone)]
 struct Backend {
@@ -22,22 +20,6 @@ struct Backend {
     api: CrateApi,
     sparse: CrateIndex,
     cache: CrateCache,
-}
-
-#[derive(Default, Debug, Clone)]
-pub struct Settings {
-    inner: Arc<RwLock<InnerSettings>>,
-}
-
-#[derive(Default, Debug, Clone, Deserialize)]
-pub struct LspSettings {
-    #[serde(rename = "useApi", default)]
-    pub use_api: Option<bool>,
-}
-
-#[derive(Default, Debug, Clone, Deserialize)]
-pub struct InnerSettings {
-    lsp: LspSettings,
 }
 
 impl Backend {
@@ -51,17 +33,8 @@ impl Backend {
             .map(|dependency| dependency.name.as_str())
             .collect();
 
-        let use_api = self
-            .settings
-            .inner
-            .read()
-            .await
-            .lsp
-            .use_api
-            .unwrap_or_default();
-
         // Get the newest version of each crate that appears in the manifest.
-        let newest_packages = if use_api {
+        let newest_packages = if self.settings.use_api().await {
             self.api
                 .fetch_versions(self.cache.clone(), &dependency_names)
                 .await
@@ -120,16 +93,9 @@ impl Backend {
 #[tower_lsp::async_trait]
 impl LanguageServer for Backend {
     async fn initialize(&self, params: InitializeParams) -> Result<InitializeResult> {
-        {
-            // Populate language server settings.
-            let settings: InnerSettings = params
-                .initialization_options
-                .and_then(|options| serde_json::from_value(options).ok())
-                .unwrap_or_default();
-
-            let mut internal_settings = self.settings.inner.write().await;
-            internal_settings.lsp.use_api = settings.lsp.use_api;
-        };
+        if let Some(settings) = params.initialization_options {
+            self.settings.populate_from(settings).await;
+        }
 
         Ok(InitializeResult {
             server_info: None,
